+++ infrastructure/files.ts
@@ -1,3 +1,4 @@
+/// <reference path="../.sst/platform/config.d.ts" />
 
 export const Files = new sst.aws.Bucket("Files", {
   access: 'cloudfront',